requirements.txt
pyodbc>=4.0.30
sqlalchemy>=1.4.0
pandas>=1.3.0
psutil>=5.9.0
python-dotenv>=0.19.0
rich>=10.0.0
graphviz>=0.17
pytest>=6.0.0  # for running tests